refactored all the code by moving them from `trainer.py` to `utils.py` and `visuals.py`
