--- src/lib/machine_learning/module/xml_parser.py
@@ -1,51 +1,0 @@
-import xml.etree.ElementTree as ET
-import glob
-import pandas as pd
-
-
-def xml_to_csv(path):
-    """Iterates through all .xml files (generated by labelImg) in a given directory and combines
-    them in a single Pandas dataframe.
-
-    Parameters:
-    ----------
-    path : str
-        The path containing the .xml files
-    Returns
-    -------
-    Pandas DataFrame
-        The produced dataframe
-    """
-
-    xml_list = []
-    for xml_file in glob.glob(path + "/*.xml"):
-        tree = ET.parse(xml_file)
-        root = tree.getroot()
-        filename = root.find("filename").text
-        width = int(root.find("size").find("width").text)
-        height = int(root.find("size").find("height").text)
-        for member in root.findall("object"):
-            bndbox = member.find("bndbox")
-            value = (
-                filename,
-                width,
-                height,
-                member.find("name").text,
-                int(bndbox.find("xmin").text),
-                int(bndbox.find("ymin").text),
-                int(bndbox.find("xmax").text),
-                int(bndbox.find("ymax").text),
-            )
-            xml_list.append(value)
-    column_name = [
-        "filename",
-        "width",
-        "height",
-        "classname",
-        "xmin",
-        "ymin",
-        "xmax",
-        "ymax",
-    ]
-    xml_df = pd.DataFrame(xml_list, columns=column_name)
-    return xml_df
